# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "LINUX DO Credit",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/api.php": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "name": "act",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "key",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "out_trade_no",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "pid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "trade_no",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/payment.QueryMerchantOrderResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "parameters": [
                    {
                        "description": "退款请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment.RefundOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/payment.RefundMerchantOrderResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/merchant/payment": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "parameters": [
                    {
                        "description": "支付订单请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment.PayOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
...
        "/api/v1/merchant/payment/order": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单号",
                        "name": "order_no",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
...
        "/api/v1/payment/transfer": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "parameters": [
                    {
                        "description": "转账请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
...
        "/pay/submit.php": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "payment.CreateOrderRequest": {
            "type": "object",
            "required": [
                "amount",
                "order_name"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "merchant_order_no": {
                    "type": "string"
                },
                "order_name": {
                    "type": "string",
                    "maxLength": 64
                },
                "payment_type": {
                    "type": "string"
                },
                "remark": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "payment.PayOrderRequest": {
            "type": "object",
            "required": [
                "order_no",
                "pay_key"
            ],
            "properties": {
                "order_no": {
                    "type": "string"
                },
                "pay_key": {
                    "type": "string",
                    "maxLength": 6
                }
            }
        },
        "payment.QueryMerchantOrderResponse": {
            "type": "object",
            "properties": {
                "addtime": {
                    "type": "string",
                    "example": "2023-12-08 12:00:00"
                },
                "code": {
                    "type": "integer",
                    "example": 1
                },
                "endtime": {
                    "type": "string",
                    "example": "2023-12-08 12:05:00"
                },
                "money": {
                    "type": "string",
                    "example": "10.00"
                },
                "msg": {
                    "type": "string",
                    "example": "查询订单号成功！"
                },
                "name": {
                    "type": "string",
                    "example": "商品名称"
                },
                "out_trade_no": {
                    "type": "string",
                    "example": "M202312080001"
                },
                "pid": {
                    "type": "string",
                    "example": "1001"
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "trade_no": {
                    "type": "string",
                    "example": "123456"
                },
                "type": {
                    "type": "string",
                    "example": "epay"
                }
            }
        },
        "payment.RefundMerchantOrderResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 1
                },
                "msg": {
                    "type": "string",
                    "example": "退款成功"
                }
            }
        },
        "payment.RefundOrderRequest": {
            "type": "object",
            "required": [
                "key",
                "money",
                "pid",
                "trade_no"
            ],
            "properties": {
                "key": {
                    "type": "string"
                },
                "money": {
                    "type": "number"
                },
                "out_trade_no": {
                    "type": "string"
                },
                "pid": {
                    "type": "string"
                },
                "trade_no": {
                    "type": "integer"
                }
            }
        },
        "payment.TransferRequest": {
            "type": "object",
            "required": [
                "amount",
                "pay_key",
                "recipient_id",
                "recipient_username"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "pay_key": {
                    "type": "string",
                    "maxLength": 6
                },
                "recipient_id": {
                    "type": "integer"
                },
                "recipient_username": {
                    "type": "string"
                },
                "remark": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
...
        "util.ResponseAny": {
            "type": "object",
            "properties": {
                "data": {},
                "error_msg": {
                    "type": "string",
                    "example": ""
                }
            }
        }
    }
}